+++ src/components/OrderPanel/OrderPanel.js
@@ -33,6 +33,7 @@ import {
   isNegotiationProcess,
   isInquiryProcess,
   isPurchaseProcess,
+  isDownloadProcess,
   resolveLatestProcessName,
 } from '../../transactions/transaction';
 
@@ -60,6 +61,10 @@ const InquiryWithoutPaymentForm = loadable(() =>
 );
 const ProductOrderForm = loadable(() =>
   import(/* webpackChunkName: "ProductOrderForm" */ './ProductOrderForm/ProductOrderForm')
+);
+
+const DigitalDownloadForm = loadable(() =>
+  import(/* webpackChunkName: "DigitalDownloadForm" */ './DigitalDownloadForm/DigitalDownloadForm')
 );
 
 const NegotiationForm = loadable(() =>
@@ -305,6 +310,8 @@ const OrderPanel = props => {
     fetchLineItemsError,
     payoutDetailsWarning,
     showListingImage,
+    hideAuthorInfo,
+    hidePrice,
   } = props;
 
   const publicData = listing?.attributes?.publicData || {};
@@ -319,7 +326,8 @@ const OrderPanel = props => {
   const isBooking = isBookingProcess(processName);
   const isPurchase = isPurchaseProcess(processName);
   const isNegotiation = isNegotiationProcess(processName);
-  const isPaymentProcess = isBooking || isPurchase || isNegotiation;
+  const isDownload = isDownloadProcess(processName);
+  const isPaymentProcess = isBooking || isPurchase || isNegotiation || isDownload;
 
   const showPriceMissing = isPaymentProcess && !isNegotiation && !price;
   const showInvalidCurrency =
@@ -350,6 +358,7 @@ const OrderPanel = props => {
   const showProductOrderForm =
     mounted && shouldHavePurchase && !isClosed && typeof currentStock === 'number';
 
+  const showDownloadForm = mounted && !isClosed && isDownload;
   const showInquiryForm = mounted && !isClosed && isInquiry;
   // if listing is a request, we show the negotiation form (reverse negotiation). User (provider) needs to make an offer first.
   const showNegotiationForm = mounted && !isClosed && isNegotiation && unitType === REQUEST;
@@ -450,24 +459,27 @@ const OrderPanel = props => {
             {subTitleText ? <div className={css.orderHelp}>{subTitleText}</div> : null}
           </div>
         )}
+        {!hidePrice && (
+          <PriceMaybe
+            price={price}
+            publicData={publicData}
+            validListingTypes={validListingTypes}
+            intl={intl}
+            marketplaceCurrency={marketplaceCurrency}
+          />
+        )}
 
-        <PriceMaybe
-          price={price}
-          publicData={publicData}
-          validListingTypes={validListingTypes}
-          intl={intl}
-          marketplaceCurrency={marketplaceCurrency}
-        />
-
-        <div className={css.author}>
-          <AvatarSmall user={author} className={css.providerAvatar} />
-          <span className={css.providerNameLinked}>
-            <FormattedMessage id="OrderPanel.author" values={{ name: authorLink }} />
-          </span>
-          <span className={css.providerNamePlain}>
-            <FormattedMessage id="OrderPanel.author" values={{ name: authorDisplayName }} />
-          </span>
-        </div>
+        {!hideAuthorInfo && (
+          <div className={css.author}>
+            <AvatarSmall user={author} className={css.providerAvatar} />
+            <span className={css.providerNameLinked}>
+              <FormattedMessage id="OrderPanel.author" values={{ name: authorLink }} />
+            </span>
+            <span className={css.providerNamePlain}>
+              <FormattedMessage id="OrderPanel.author" values={{ name: authorDisplayName }} />
+            </span>
+          </div>
+        )}
 
         {showPriceMissing ? (
           <PriceMissing />
@@ -529,6 +541,13 @@ const OrderPanel = props => {
             shippingEnabled={shippingEnabled && displayShipping}
             displayDeliveryMethod={displayPickup || displayShipping}
             onContactUser={onContactUser}
+            {...sharedProps}
+          />
+        ) : showDownloadForm ? (
+          <DigitalDownloadForm
+            formId="OrderPanelDigitalDownloadForm"
+            finePrintComponent={SubmitFinePrint}
+            isOwnListing={isOwnListing}
             {...sharedProps}
           />
         ) : showInquiryForm ? (
@@ -597,6 +616,8 @@ const OrderPanel = props => {
               <FormattedMessage id="OrderPanel.ctaButtonMessageMakeOffer" />
             ) : showRequestQuoteForm ? (
               <FormattedMessage id="OrderPanel.ctaButtonMessageRequestAQuote" />
+            ) : showDownloadForm ? (
+              <FormattedMessage id="OrderPanel.ctaButtonMessageDownload" />
             ) : (
               <FormattedMessage id="OrderPanel.ctaButtonMessageInquiry" />
             )}