+++ src/containers/TransactionPage/TransactionPage.stateDataDownload.js
@@ -1,0 +1,92 @@
+import {
+  TX_TRANSITION_ACTOR_CUSTOMER as CUSTOMER,
+  TX_TRANSITION_ACTOR_PROVIDER as PROVIDER,
+  CONDITIONAL_RESOLVER_WILDCARD,
+  ConditionalResolver,
+} from '../../transactions/transaction';
+import { transitions, states } from '../../transactions/transactionProcessDownload';
+
+/**
+ * Get state data against digital download process for the TransactionPage's UI.
+ * I.e. info about showing action buttons, current state etc.
+ *
+ * @param {*} txInfo details about the specific transaction
+ * @param {*} processInfo  details about the transaction process
+ */
+export const getStateDataForDownloadProcess = (txInfo, processInfo) => {
+  const { transaction, transactionRole, nextTransitions } = txInfo;
+  const isProviderBanned = transaction?.provider?.attributes?.banned;
+  const _ = CONDITIONAL_RESOLVER_WILDCARD;
+
+  const {
+    processName,
+    processState,
+    isCustomer,
+    actionButtonProps,
+    leaveReviewProps,
+  } = processInfo;
+
+  return (
+    new ConditionalResolver([processState, transactionRole])
+      .cond([states.INQUIRY, CUSTOMER], () => {
+        const transitionNames = Array.isArray(nextTransitions)
+          ? nextTransitions.map(t => t.attributes.name)
+          : [];
+        const requestAfterInquiry = transitions.REQUEST_PAYMENT_AFTER_INQUIRY;
+        const hasCorrectNextTransition = transitionNames.includes(requestAfterInquiry);
+        const showOrderPanel = !isProviderBanned && hasCorrectNextTransition;
+        return { processName, processState, showOrderPanel, showDetailCardHeadings: true };
+      })
+      .cond([states.INQUIRY, PROVIDER], () => {
+        return { processName, processState, showDetailCardHeadings: true };
+      })
+      .cond([states.PURCHASED, PROVIDER], () => {
+        return {
+          processName,
+          processState,
+          showDetailCardHeadings: true,
+          showExtraInfo: true,
+        };
+      })
+      .cond([states.PURCHASED, CUSTOMER], () => {
+        return {
+          processName,
+          processState,
+          showDetailCardHeadings: true,
+          showReport: true,
+          showExtraInfo: true,
+        };
+      })
+      .cond([states.REPORTED, CUSTOMER], () => {
+        return { processName, processState, showDetailCardHeadings: true, showExtraInfo: true };
+      })
+      .cond([states.REPORTED, PROVIDER], () => {
+        return { processName, processState, showDetailCardHeadings: true, showExtraInfo: true };
+      })
+      .cond([states.COMPLETED, PROVIDER], () => {
+        return { processName, processState, showDetailCardHeadings: true };
+      })
+      .cond([states.COMPLETED, CUSTOMER], () => {
+        return {
+          processName,
+          processState,
+          showDetailCardHeadings: true,
+          showExtraInfo: true,
+          showReviewAsFirstLink: true,
+          showActionButtons: true,
+          primaryButtonProps: leaveReviewProps,
+        };
+      })
+      .cond([states.REVIEWED, _], () => {
+        return { processName, processState, showDetailCardHeadings: true, showReviews: true };
+      })
+      .cond([states.CANCELED, _], () => {
+        return { processName, processState, showDetailCardHeadings: true, showExtraInfo: true };
+      })
+      // PENDING_PAYMENT and PENDING_EXPIRED are handled by default
+      .default(() => {
+        return { processName, processState, showDetailCardHeadings: true };
+      })
+      .resolve()
+  );
+};